Set up webpack for bundling and transpilation.
