+++ .gitignore
@@ -5,5 +5,5 @@ secrets*.js
 .vscode
 .sass-cache
 
-public/build/css/*
+public/build/**
 !.gitkeep
