+++ app/views/campaign.hbs
@@ -17,7 +17,7 @@
             <h5 class="section-header">Your Congressperson is</h5>
             <h3 class="rep-name">{[{repInfo.title}]} {[{repInfo.first_name}]} {[{repInfo.last_name}]} ({[{repInfo.party}]})</h3>
             <img
-              src="https://theunitedstates.io/images/congress/450x550/{[{repInfo.bioguide_id}]}.jpg"
+              ng-src="https://theunitedstates.io/images/congress/450x550/{[{repInfo.bioguide_id}]}.jpg"
               class="rep-img center-block"
               alt="Your Rep"
             />
