+++ app/views/layouts/main.hbs
@@ -17,24 +17,29 @@
 
     <link href="/build/css/styles.css" rel="stylesheet">
 
-    <!-- Just for debugging purposes. Don't actually copy these 2 lines! -->
-    <!--[if lt IE 9]><script src="../../assets/js/ie8-responsive-file-warning.js"></script><![endif]-->
-    <script src="../../assets/js/ie-emulation-modes-warning.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
     <script src="http://maps.googleapis.com/maps/api/js?key=AIzaSyDM2H_5tEaB4OM1Dr2iR893KmwczVNXExs&libraries=places" type="text/javascript"></script>
-
-    <script src="/js/vendor/bundle.js"></script>
-    <script src="/js/core.js"></script>
+    <script src="/build/js/vendor.bundle.js"></script>
 
     <script>
-      function openNav() {
-          document.getElementById("sideMenu").style.width = "320px";
-      }
+        function openNav() {
+            document.getElementById("sideMenu").style.width = "320px";
+        }
 
-      function closeNav() {
-          document.getElementById("sideMenu").style.width = "0";
-      }
+        function closeNav() {
+            document.getElementById("sideMenu").style.width = "0";
+        }
+
+        angular.module('helloGov', ['ngMapAutocomplete', 'ngclipboard'])
+            .config(function($interpolateProvider) {
+                $interpolateProvider.startSymbol('{[{').endSymbol('}]}');
+            })
+            .constant('constants', {
+                API_ROOT: '/api'
+            });
+
     </script>
+    <script src="/build/js/app.bundle.js"></script>
   </head>
 
   <body class="{{{body_class}}}">
@@ -65,17 +70,5 @@
   </body>
 
   <footer>
-    <script src="/js/vendor/handlebars.runtime.min.js"></script>
-    <script>
-        (function () {
-            var revive    = Handlebars.template,
-                templates = Handlebars.templates = Handlebars.templates || {};
-
-          {{#templates}}
-            templates['{{{name}}}'] = revive({{{template}}});
-          {{/templates}}
-        }());
-    </script>
-
   </footer>
 </html>
