+++ package.json
@@ -6,12 +6,15 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node index.js",
-    "build": "NODE_ENV=production sass public/scss/styles.scss public/build/css/styles.css",
+    "build": "npm run build-css && npm run build-js",
+    "build-css": "NODE_ENV=production sass public/scss/styles.scss public/build/css/styles.css",
+    "build-js": "NODE_ENV=production webpack",
     "lint": "eslint -c .eslintrc.js --fix ."
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "angular": "^1.6.4",
     "body-parser": "^1.15.2",
     "clipboard": "^1.6.0",
     "connect-mongo": "^1.3.2",
@@ -40,12 +43,20 @@
     "ua-parser": "^0.3.5"
   },
   "devDependencies": {
+    "babel-cli": "^6.24.0",
+    "babel-loader": "^6.4.1",
+    "babel-preset-env": "^1.3.2",
+    "clean-webpack-plugin": "^0.1.16",
     "eslint": "^3.18.0",
     "eslint-config-standard": "^7.1.0",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-node": "^4.2.1",
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-standard": "^2.1.1",
-    "node-sass-middleware": "^0.11.0"
+    "ng-annotate-webpack-plugin": "^0.2.1-pre",
+    "node-sass-middleware": "^0.11.0",
+    "uglifyjs-webpack-plugin": "^0.4.1",
+    "webpack": "^2.3.2",
+    "webpack-dev-middleware": "^1.10.1"
   }
 }
