+++ public/js/index.js
@@ -1,0 +1,11 @@
+// TODO: Remove this file if/when we are truly using JS modules properly
+
+require('./admin.controller');
+require('./analytics.controller');
+require('./campaign.controller');
+require('./matchInput.directive');
+require('./signup.controller');
+require('./success.controller');
+require('./user.controller');
+require('./visitor.controller');
+
